feat(debug): Add diagnostic logging to 3D view
Adds extensive logging to the ElevationView3D and RouteRenderer classes to help diagnose a visual issue where the 3D scene is not appearing.

This logging will trace:
- The number of vertices and indices generated for the route mesh.
- The calculated bounding box of the route.
- The final camera position and view center.
