+++ src/build_info.cpp
@@ -1,0 +1,5 @@
+#include "build_info.h"
+#include <QString>
+
+// Define build timestamp for version tracking
+const QString BUILD_TIMESTAMP = __DATE__ " " __TIME__;
